Move generators into a dedicated namespace
